+++ src/include/sensors.h
@@ -9,12 +9,12 @@
  * generates random data and assigns it to a double *
  * 
  */
-int test_sensor(double *ptr, mraa_context context);
+int test_sensor(float *ptr, mraa_context context);
 
 /**
  * reads and calibrates data from the FPS V2 Range 5V Pressure Sensor 
  * 
  */
-int fps_v2_range_5v(double *ptr, mraa_context context);
+int fps_v2_range_5v(float *ptr, mraa_context context);
 
 #endif
